Add files via upload
The first issue that came with transposing Claire De Lune was the time signature was wrong. It read the 9 on 9/8 as 6l. It also hade issues when I would fix the timing or rhythm, it would push more issues into the next bar. By bar 10, it was a chaotic mess. I am still trying to understand the double voice but it would move that around too.